BasicDemo/requirements.txt
cnocr==2.2.2.3
cnstd==1.2.2
loguru==0.7.0
numpy==1.23.5
opencv_contrib_python==4.5.4.60
opencv_python==4.7.0.72
opencv_python_headless==4.5.4.60
pandas==2.0.1
Pillow==10.0.0
pyserial==3.4
pywin32==306
wmi==1.5.1
